Don't have output if unnecessary; never overwrite local files; ensure file exists before trying to unzip it; remove pack.properties.

+++ releng/org.eclipse.linuxtools.releng/tools/linuxtoolsupdatesite-trunk.sh
@@ -12,9 +12,7 @@ rsync -a ${dropLocation} ${downloadsDir}
 
 timestamp=$(date +"%Y-%M-%d %H:%M:%S")
 
-echo "Starting creation of update site:  ${timestamp}"
-
-pushd ${updateSite}
+cd ${updateSite}
 
 if [ ! -e site.xml ]; then
   svn export http://dev.eclipse.org/svnroot/technology/org.eclipse.linuxtools/releng/branches/0.2.0/org.eclipse.linuxtools.updatesite/updates-nightly/site.xml
@@ -26,16 +24,18 @@ fi
 # 20 is for 2009, 2010, etc.
 latest=$(ls -1 ${downloadsDir} | grep ^[NS]20 | cut -c2- | sort | tail -n 1)
 
+if [ -e ${downloadsDir}/*${latest}/linuxtools-Master*.zip ]; then
+
 # Clean out old
 #rm -f old.tar.bz2
 #tar jcf old.tar.bz2 *
-rm -rf features plugins
+rm -rf features plugins pack.properties
 
-unzip -q ${downloadsDir}/*${latest}/linuxtools-Master*.zip
+unzip -q -n ${downloadsDir}/*${latest}/linuxtools-Master*.zip
 mv eclipse/* .
 rmdir eclipse
 
 # Update site.xml with the versions we have in this build
 for f in `ls features/*.jar`; do version=$(echo $f | sed -e 's:features/.*_::' -e 's:\.jar::'); name=$(echo $f | sed -e 's:features/::' -e 's:\.jar::' -e 's:_::' -e "s:$version::"); sed -i -e "/${name}/ s/_.*\.jar/_${version}\.jar/" -e "/${name}/ s/version=\".*\"/version=\"${version}\"/" site.xml; done
 
-popd
+fi
